Stack task. Tests and docs.
I did a little change of stack realization. Added private method of
cleaning Stack to use it in destructor and assignment operator. Replaced
some code into macros. Added new namespace in mainfile with realization
of some tests of Stack. Changed comments to doctype style for Doxygen
and created documentation with it.